Get weekdays weather from city search
One functionality of this app is to get and display the overview of the
weather for each day of the week. This commit handle the part of getting and
creating the weather object for each day that will be later displayed.

User search for a city. Geocoding API gets coords of city. Coords
are passed to weather API which gets the weather of each day od the week
in that city. Not all datas are needed, new objects for each with important
parameters are created and stored in array.

Still trying to understand how to transfer datas from the end of creation
to the display part.