Add UI day specifics on click
Every time that the user click on a day in the overview the specifics
of that day must be displayed.

A click event listener has been placed on each day in the overview.
When user click on a day a function is called with the values of that
day as an argument. The function uses already created input fields and fill
the text content with the data coming from the day we clicked.