Display UI current position
When use log into the application the weather of his current position must be
displayed and loaded.

When user search for a country our call API gets his coords.
But to get the current position of the user we use use the JS geolocation API
which return an object with the city's coords. We first reverse geocode the
coords to get the name of the current city, and then use the same logic when
user search for a city.